+++ include/axl/axl-version.h
@@ -26,8 +26,8 @@
 
 #define AXL_VERSION_MAJOR   2
 #define AXL_VERSION_MINOR   3
-#define AXL_VERSION_PATCH   0
-#define AXL_VERSION_STRING  "2.3.0"
+#define AXL_VERSION_PATCH   1
+#define AXL_VERSION_STRING  "2.3.1"
 
 /* Encoded as 0xMMmmpp (major, minor, patch) for simple comparisons. */
 #define AXL_VERSION_NUMBER \